refactored dbs into their own file
