+++ .github/CODEOWNERS
@@ -34,6 +34,7 @@ src/roles.js @WilsontheWolf
 src/util.js @WilsontheWolf @KommentatorForAll
 src/xpHandling.js @KommentatorForAll
 src/index.js @WilsontheWolf @KommentatorForAll @Chicken
+src/dbs.js @WilsontheWolf
 src/helper/ @WilsontheWolf
 
 # Config
